feat!: move fastq functionality to reads subcommand
